test(setup): address review feedback on pin orchestration tests
